feat: Sevan SQL repository

+++ src/init.sh
@@ -13,16 +13,21 @@ if [ -d /code/src/storage/repository_plugins ]; then
   echo "Checking for plugin requirements..."
   cd /code/src/storage/repository_plugins
   PLUGINS=$(ls)
-  for plugin in $PLUGINS ; do
-      echo "Installing requirements for plugin $plugin"
-      pip install -r "$plugin/requirements.txt"
-  done
+    for plugin in $PLUGINS ; do
+      if [ -f "$plugin/requirements.txt" ]; then
+          echo "Installing requirements for plugin $plugin"
+          pip install -r "$plugin/requirements.txt"
+        else
+            echo "No requirements.txt found for plugin $plugin"
+      fi
+    done
 
   cd -
 fi
 
 
 envsubst < /code/src/system_DS_template.json > /tmp/DMSS_systemDS.json
+envsubst < /code/src/sql_system_DS_template.json > /tmp/SQL_systemDS.json
 
 if [ "$1" = 'api' ]; then
   if [ "${RESET_DATA_SOURCE:-"on"}" == "on" ]; then
